Add  some styles
